Adding -x option
1.6.0-1.6.0e fixes
Fixed an issue when displaying dispatch date when there are multiple
disks recoveries on the same node.
Fixed an issue with days displaying higher number then when the disk has
been failed out previously.
Fixed an issue with gen 1 internal drives not being able to be assigned
and enhanced output for gen1/2 internal disks
Fixed an issue where it was reporting garbage mds disks when the system
is setup with mixed disks
Fixed an issue where it was producing the same disk as mixed and SS
More information in the .info file for historical purposes
1.6.1 add -x option for xfs_repair, currently only support for BEATLE